+++ src/gui/callbacks.h
@@ -17,7 +17,7 @@ void set_display_mode();
 void adjust_exclude(int n, gboolean changed);
 int adjust_sellabel();
 void set_GUI_CWD();
-void set_GUI_MEM(unsigned long size);
+void set_GUI_MEM(unsigned long long size);
 void set_GUI_DiskSpace(int64_t mem);
 void set_GUI_misc();
 void set_icon_entry(GtkEntry *entry, gchar *string);
